# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Clinical Trial Data Collection Platform API
  description: |
    HIPAA-compliant RESTful API for clinical trial patient data collection platform.
    
    This API provides secure endpoints for:
    - User authentication with JWT and MFA support
    - Patient data management with column-level encryption
    - Questionnaire creation and assignment
    - Secure response collection and storage
    - Media file uploads with virus scanning
    - Comprehensive audit logging
    
    ## Security Features
    - JWT Bearer token authentication
    - Role-based access control (Physician vs Patient)
    - Multi-factor authentication (MFA) support
    - Column-level data encryption for PHI
    - HIPAA-compliant audit logging
    - Rate limiting and request throttling
    - Virus scanning for file uploads
    
    ## Data Privacy
    - All personally identifiable information (PII) is encrypted at rest
    - PHI is never exposed in error messages
    - Comprehensive audit trails for all data access
    - Secure token-based session management
    
    ## Authentication Flow
    1. Login with email/password to receive access token
    2. Include Bearer token in Authorization header for all requests
    3. Refresh tokens before expiration using /auth/refresh
    4. Complete MFA verification if enabled
    
  version: 1.0.0
  contact:
    name: Clinical Trial Platform Development Team
    email: dev@clinicaltrial.platform
  license:
    name: Proprietary
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token obtained from /auth/login endpoint.
        
        Example: Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Authentication required
        code:
          type: string
          description: Machine-readable error code
          example: AUTH_REQUIRED
        details:
          type: array
          items:
            type: object
          description: Additional error details (validation errors, etc.)
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        requestId:
          type: string
          description: Unique request identifier for debugging
      required:
        - error
        - code
    PaginationResponse:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Items per page
          example: 10
        total:
          type: integer
          minimum: 0
          description: Total number of items
          example: 25
        pages:
          type: integer
          minimum: 1
          description: Total number of pages
          example: 3
      required:
        - page
        - limit
        - total
        - pages
paths:
...
  /physicians/dashboard:
    get:
      tags:
        - Physicians
      summary: Get physician dashboard data
      description: |
        Retrieve comprehensive dashboard information for the authenticated physician.
        
        Includes:
        - Patient count under care
        - Active studies participation
        - Pending responses requiring review
        - Recent patient activity
        
        Data is filtered to only include patients under the physician's care.
      responses:
        '200':
          description: Dashboard data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  dashboard:
                    type: object
                    properties:
                      overview:
                        type: object
                        properties:
                          totalPatients:
                            type: integer
                            description: Number of patients under care
                            example: 25
                          activeStudies:
                            type: integer
                            description: Number of active studies physician is involved in
                            example: 3
                          pendingResponses:
                            type: integer
                            description: Number of responses awaiting review
                            example: 8
                      recentActivity:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            type:
                              type: string
                              example: response
                            title:
                              type: string
                              example: Daily Symptom Assessment
                            patientId:
                              type: string
                              description: Anonymized patient identifier
                            status:
                              type: string
                              example: completed
                            timestamp:
                              type: string
                              format: date-time
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Physician access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /physicians/dashboard/stats:
    get:
      tags:
        - Physicians
      summary: Get physician dashboard statistics
      description: |
        Retrieve detailed statistics for the physician dashboard with time-based filtering.
        
        Provides metrics such as:
        - Active patient count in specified timeframe
        - Response completion rates
        - Average completion times
        - Response status breakdown
      parameters:
        - name: timeframe
          in: query
          description: Number of days to include in statistics
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 30
          example: 30
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stats:
                    type: object
                    properties:
                      timeframe:
                        type: string
                        example: "30 days"
                      activePatients:
                        type: integer
                        example: 18
                      totalResponses:
                        type: integer
                        example: 142
                      completedResponses:
                        type: integer
                        example: 128
                      inProgressResponses:
                        type: integer
                        example: 14
                      avgCompletionTime:
                        type: integer
                        description: Average completion time in minutes
                        example: 12
  /physicians/profile:
    get:
      tags:
        - Physicians
      summary: Get physician profile
      description: |
        Retrieve the complete profile information for the authenticated physician.
        
        Includes professional details, organization affiliation, and account settings.
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  profile:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                        format: email
                      firstName:
                        type: string
                        example: John
                      lastName:
                        type: string
                        example: Smith
                      title:
                        type: string
                        example: Dr.
                      specialization:
                        type: string
                        example: Cardiology
                      licenseNumber:
                        type: string
                        example: MD123456
                      phone:
                        type: string
                        example: +1-555-0123
                      department:
                        type: string
                        example: Internal Medicine
                      role:
                        type: string
                        example: attending_physician
                      permissions:
                        type: array
                        items:
                          type: string
                        example: ["create_questionnaires", "review_responses"]
                      organization:
                        type: object
                        properties:
                          name:
                            type: string
                            example: Central Medical Center
                          type:
                            type: string
                            example: hospital
    put:
      tags:
        - Physicians
      summary: Update physician profile
      description: |
        Update the physician's profile information.
        
        Only the provided fields will be updated. Sensitive fields like email and 
        license number may require additional verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 50
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 50
                title:
                  type: string
                  maxLength: 20
                specialization:
                  type: string
                  maxLength: 100
                licenseNumber:
                  type: string
                  maxLength: 50
                phone:
                  type: string
                  pattern: '^\+?[1-9]\d{1,14}$'
                department:
                  type: string
                  maxLength: 100
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Profile updated successfully
                  profile:
                    type: object
                    description: Updated profile data
  /physicians/patients:
    get:
      tags:
        - Physicians
      summary: Get patients under physician's care
      description: |
        Retrieve a paginated list of patients under the physician's care.
        
        Results can be filtered by patient status and searched by patient ID or email.
        Patient data is appropriately anonymized for HIPAA compliance.
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Search by patient ID or email
          schema:
            type: string
            maxLength: 100
        - name: status
          in: query
          description: Filter by patient status
          schema:
            type: string
            enum: [active, inactive, completed, withdrawn]
      responses:
        '200':
          description: Patients retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  patients:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        email:
                          type: string
                          format: email
                        patientId:
                          type: string
                          description: Unique patient identifier
                        status:
                          type: string
                          enum: [active, inactive, completed, withdrawn]
                        enrollmentDate:
                          type: string
                          format: date
                        relationshipType:
                          type: string
                          example: primary_care
                        relationshipStarted:
                          type: string
                          format: date-time
                        totalResponses:
                          type: integer
                          example: 15
                        completedResponses:
                          type: integer
                          example: 12
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'

  # Patient Endpoints
...
  /patients/questionnaires/{questionnaireId}:
    get:
      tags:
        - Patients
      summary: Get questionnaire details with patient's response
      description: |
        Retrieve detailed questionnaire information including questions and any existing
        patient response data.
        
        Response data is decrypted and returned if the patient has started the questionnaire.
        Access permissions are verified before returning data.
      parameters:
        - name: questionnaireId
          in: path
          required: true
          description: Unique questionnaire identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Questionnaire details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  questionnaire:
                    allOf:
                      - $ref: '#/components/schemas/Questionnaire'
                      - type: object
                        properties:
                          response:
                            allOf:
                              - $ref: '#/components/schemas/PatientResponse'
                              - type: object
                                description: Current patient response (null if not started)
        '403':
          description: Access denied to questionnaire
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error: Access denied
                message: You do not have access to this questionnaire
                code: QUESTIONNAIRE_ACCESS_DENIED
        '404':
          description: Questionnaire not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...